async: update hooks for use in celery
The global variable 'conf' is not needed anymore, because the
configuration is passed to the celery tasks as arguments.

The 'session' argument has been removed from add_package/rm_package,
because a sqlalchemy session can't be serialized and passed to a celery
task. Instead, each plugin must import the session from
new_updater.celery, which will be setup on each worker (using the
'celeryd_init' signal).